avoid checking rakudo-specific error
I could not use X:: here as it still throws X::AdHoc.
I also could not switch to eval_dies_ok as evaluated code has "die" in
it - it will be impossible to say if code died to compilation failure
or die inside caused it.
So I reversed test logic and I'm checking if die in evaluated code was
not executed by comparing $! value after eval.