Add simple test of random number distribution.
Rakudo JVM currently has a bug where rand_I is biased towards numbers at the
beginning of a range. This simple check should detect that by rolling a number
from 0 to 4, then summing up the number of times 3 or 4 was found. In a proper
random number generator, this should be on the order of 400. We check to see
if it is less than 300 and complain if it is.

This might happened by change to a good random number generator occasionally,
of course.