multichar line sep on server socket
This test verifies that the sockets can operate on
multichar line seperators for calls to get.

This should already have been verified by test 5.
Somehow that test passes when this one does not.
Test 5 seems to the on the client socket.

jnthn mentioned that MOAR does not like multichar
line sepearators. But the problem seems to persist
for Parrot and JVM backend in a slightly different
manner.

I only tested this patch on Linux.