Add a working version of RT#121789.
Arguably, fixing the test doesn't fix the ticket, but my
understanding of what the snippet is supposed to do is
getting the matched integers, which a raw .list call would not.
Calling .list on a $/ obtained from :g should list the submatches
for each repetition of the match, and not the captures, similarly
to how calling .list on a 'normal' $/ gives the matches and not
their captures.