Simplify the Proc::Async.kill tests
Unfortunately, the actual kill can not be done here, as it causes the
current process to lose the remote process, so the P::A promise is never
kept or broken.  This even affects it being a part of an anyof with a timeout,
so we cannot make the test continue.  :-(   Hope this will help jnthn find
what the heck is going on.