Tests for aliasing different kinds of native.
This covers a bug found in Inline::Perl5 where a lexref to an int32
didn't work out. However, the exact case in Inline::Perl5 is not an
ideal test case as a slightly smarter optimizer will elide the lexref.
Writing a test like this, however, will be a bit harder on it.