Replace Captures by itemization for getting references in P5
Captures are listy, while references in Perl 5 are the exact opposite:
a way to turn compound objects into scalars. Using Captures to get
references when passing objects like arrays and hashes to Perl 5 code
therefore does not make any sense.

Itemization on the other hand matches Perl 5's semantics much better.
Itemizing an array when passing it to a Perl 5 function and having it
arrive as an arrayref seems like a very natural way to do it.