Script to generate some mass NFG tests.
We look through the Unicode Normalization Tests for cases where the
NFC form contains non-starters, giving us some good candidates for
NFG tests. We then generate tests checking we get the correct results
for .chars, and that we can get back from NFG to codepoints. Mostly we
will be going to NFC, so we generate more tests for that, but produce
100 each for the other forms too just to quickly sanity test them.