Make a test robust in the face of NFG.
It had a hash of opener/closer pairs. However, in an implementation
that does NFG, due to singleton decompositions in Unicode, two of the
opener/closer pairs ended up equal after normalization, resulting in a
hash with an element less and thus a test less being run. Use a list
of pairs instead.