fix flapping(?) test. indent cleanup
Passes all tests, including supposed previously flapping test.

Convert mix of tabs+spaces to all space indentation. 
Corrected the binary server code to actually work.
Passes previously flapping test RT 122318? (Not sure if this is fixed or not. The RT has 0 information so its impossible to know)

I was not able to reproduce a flapping test before or after this PR so you may wish to un-skip the final test and close the ticket. Before this PR, it was simply FAILING the final test after some 10k iterations.
    
    nickl@localhost:~/perl6/rakudo$ prove -v -e './perl6 -Ilib' t/spec/S32-io/IO-Socket-Async.t
    t/spec/S32-io/IO-Socket-Async.t ..
    1..6
    ok 1 - Async listen on bogus hostname
    ok 2 - Async connect to unavailable server breaks promise
    ok 3 - Async connect to available server keeps promise
    ok 4 - Echo server
    ok 5 - Discard server
    ok 6 - bytes_supply
    ok