Adjust dubious tests, per RT#125815 and due to GLRification
When I GLRified :nth and :x I made a couple changes to behavior
(sorry).  They were justified: the original code intended to
die on :nth(-1) but for some reason did not (RT#125815). Now it does.

Also, the * handling was applying 0-based semantics to a 1-based
value, so that was fixed.  Added a skipped test for a lone * which
is NYI on both nom and GLR.

Left a failing test that found one thing broken during
GLRification (s:0x// for some reason still substs.)