Fix S02-literals/numeric.t
Some of the tests for different forms of Rat literals did so in a way
that calls will call .Numeric on Str (as well as any other non-Numeric
argument). Since a Rat literal that's not parsed correctly will be
interpreted as a Str, then as long as Str.Numeric can handle the given
Rat form any issues with Rat literals for those forms were hidden. These
tests were changed to ensure we don't numify any strings.

Also updated some of the tests to use the new 'does-ok' to check roles,
and replaced a smartmatch against Str with an 'isa-ok' check.