Correct «» interpolation test.
There's no evidence in the synopses that interpolated variables in «»
lists should ever be not-flat, and with the new allomorphic-based
implementation of these things in rakudo, they are in fact always a flat
list. For what it's worth the glr branch of rakudo does this too.

So, the test got corrected instead of the implementation, unless/until
evidence for the old behavior is given.