Fix tests relying on ''.Numeric -> 0
This is a really odd and surprising thing I found after switching
Str.Numeric to using val(), since val() doesn't return an Int of 0 for
the empty string.

A quick grep of the specs found no proof that this should be, and it
seems quite unintuitive. Maybe if you know P5 this makes sense, but for
me (and hopefully most others), this behavior is wrong. (At least, I
don't see any stringified numbers in that string. Do you?)