Modify test to run much faster
For some reason, the new implementation of val() and such has caused the
affected test to run **much** slower than it did before. This change
brings the test from 285 wallclock secs to 14 wallclock secs. I
unfortunately have no idea what caused the slowdown, or if the lack of
explicit stringy conversion on Methods was integral to this test.

For the time being, I figured ridding a test of unreasonable slowdown
was the best course of action.