Fudged tests of numifying certain 'slash fractionals'
These tests all make sure that you can't numify slash fractional strings
with negative denominators and/or non-integral components. Since it's
unclear what val()'s stance should be on these things.

The S02 tests are fudged because val processing takes over on a failed
Rat literal, and produces a RatStr as it stands. It's not specified if
this is expected, or if Rat/Complex literals are meant to be
all-or-nothing.