Fix tests for upcoming NFG semantic changes.
We stick in a control char to force a grapheme extender to stand alone
and so be available for matching in isolation. Note that these tests
are rather contrived, since they pair together graphemes from multiple
(natural) languages. Another possible change would have been to pick a
non-extending character in place of the extending ones for the tests
in question, but this change was the less effort.