Fix Windows line ending workarounds.
The completion of the NFG work making \r\n a single grapheme, per the
Unicode spec, means that the approaches these tests took to work on
Windows became bogus. Note that we'll probably have to take a pass
through these all again in the near future, when \n in a string will
become \r\n on Windows by default (hopefully clearing up a lot of
this cruft).