Fix samemark test
In light of the comment TimToady made in commit message fc4bf6f9d43cf3a82 :

I decided that repeating novelty combiners will likely be the more common
use of the construct than trying to tweak words from natural languages.

And the fact the test was written *before* that, I assumed the combiner
was to be repeated, rather than ignored for subsequent characters.