We can no longer .done a Supply
We can only .done a Supplier.  But the Supplier is hidden in
IO::Notification.watch-path.  So I can only hope that closing the final tap
on the Supply given by IO::Notification.watch-path *will* actually
.done the Supplier.

This unbreaks this test-file on OS X.  Other OS's are currently *not* being
tested: tests for other platforms welcome!