Fix Win32 specific IO::Spec tests so they pass
I filed a confused [bug report][1] thinking these tests were
not passing because of a failure of IO::Spec to delegate to
IO::Spec::Win32. Later, I realized that invoking methods on
IO::Spec is wrong, and one must use $*SPEC. This chunk of the
the test file probably did not get exercised much.