Skip strange adverb tests to be caught by MMD
For some yet unknown reason, using a strange adverb that should be
caught by MMD, does not always work.  Worse, if you skip the first
test that failed to die, it will not die on the next test.  So for
now, all such tests are skipped to provide a reliable test result
for the 6.c release