Fix faulty test
When in a leap second, adding a day should not move your forward to
next day, but will put you in the last second of the next day.  Just
as moving one month from Jan 31 will not put you in March, but in
the last day of February.  This incorrect behaviour was fixed post
2015.12: this test was incorrect, and is now fixed.