Mark now failing test as TODO
Because Parameter.ACCEPTS also includes post constraint checking since
bb0eab24f7e8a2f1 and having a Callable as a post constraint will always
fail the smart match (because we can not introspect the Callable), and
the signature on a Callable ( like :(&foo:() ) is compiled as a Callable,
this test no longer succeeds.

Discussion at http://irclog.perlgeek.de/perl6/2016-03-07#i_12149245

Since the idea posted there would be to store the signature directly in
the post constraint, I've todoed this test for better times.