Remove useless and actually harmful 'use lib' statements from modules.
There's no need to use lib 't/spec/packages' in a module that _is_ loaded
from that path already. If t/spec/packages was not in the repo list, we could
not have loaded the modules in the first place.

It's actually harmful as we don't allow a "use lib" in a precompiled module.
This only made it into 6.c as other restrictions prevented us from
precompiling these modules but those restrictions were an implementation
detail, not necessitated by the language design.