Correct test that did unprotected array writes.
It's only safe to grab an *already vivified* element of an array and
write to it in a cross-thread way. Since @out's elements were never
vivified, this led to concurrent binds - and thus resizes - to the
array, and thus incorrect results and crashes.

The SEGV that sometimes happened needs addressing too; that already
has an RT issue and will be covered by another test that actually
wants to create that condition, rather than this one which did it by
accident.