Send \n and end of chars echo server message.
The \n serves as a normalization terminator - that is, it means we
know that even if a combining char comes next, it will end up as an
isolate rather than forming a grapheme with the \n. A `z` could be
followed by a combiner in the next packet, so after upcoming fixes
to async socket decoding we'll now not hand such a character on
until the next packet arrives and we see there is no combiner.
This would break the test, which did not account for such NFG
behavior.