Remove mixed-level classify tests
Mixed level functionality was added to Rakudo *today* in [^1] and removed
same day in [^2], because it can't work perfectly due to key clashes (and
speculation advises for same-level multi-level classifications only). The
multi-level classify tests were added at the same time the feature was
added to Rakudo, so it's fine to remove them as they were not part of 6.c
and exited for less than 24 hours.

[1] https://github.com/rakudo/rakudo/commit/b4a4b604945772a828e9edb0fa8b1
[2] https://github.com/rakudo/rakudo/commit/74ca01d94ef8d7c25f776e6adbc27