Replace deprecated is_approx with is-approx
is_approx will reach the end of its deprecation cycle on Dec. 17. Replace it
with the new version, the is-approx. The new is-approx without explicit
tolerance provided behaves exactly as old is_approx as of today[^1], so this
commit makes no functional changes and all tests function the same.

Method used:
1) find . -iname '*' -exec perl -pi -e \
    's/(^|\s)is_approx(\(|\s)/${1}is-approx${2}/g' {} \;
2) Check manually to ensure all substitutions are done on is_approx
    without custom tolerances.
3) ???
4) Profit!

[1] https://github.com/rakudo/rakudo/commit/82432a4581d06213990e97a8e8a094444d9991d5