Make test for output of backtrace less specific
After the recent changes to backtrace handling this test
started to fail on rakudo-j. There are references to lines
in gen/jvm/CORE.setting which are causing the test to fail
because of the 's' in 'setting'.

Allowing additional lines in the backtrace still fits the test
description 'Thrown Failure outputs dual backtraces'.