Remove mass-chars.t
The script this test was generated assumed:
  The number of chars we expect is:
  - Number of codes minus number of non-starters if the test
    case begins with a starter
  - Number of codes minus number of non-starts plus one
    otherwise (to account for isolated non-starters).

This assumption did not take all the Unicode rules into account and
since more recent Unicode updates, they no longer hold true for the file
it was generated from (NormalizationTest.txt).

This test is superceeded by the recently added GraphemeBreakTest.t
which is generated from Unicode's GraphemeBreakTest.txt, which was designed
for what we seemed to be trying to test for in mass-chars.t.