Fix incorrect combinations() tests
The tests were added[^1][^2] on July 4, 2016 and are not part of 6.c.
Three of them assume the result of picking a 2-item combination from an empty list is
an 1-item list containing empty list. It was a now-fixed[^3] bug. Fourth one assumes
negative k is equivalent to k=0 and was passing only due to the test using `is` (it was
changed[^4] to `is-deeply` in a commit previous to this one).

[1] https://github.com/perl6/roast/commit/2f8f0a4ceb5a3f7e6a2c6f3216a2a9e2c70cff3d
[2] https://github.com/perl6/roast/commit/0ff33b312bbaf63dae0c19fdfbbbfae97ef53883
[3] https://github.com/rakudo/rakudo/commit/d86c419623567a7b206008323f5
[4] https://github.com/perl6/roast/commit/3ea6fa6c71