Restructure file to avoid triggering SPESH bug
This commit makes no changes to tests themselves. It merely
avoids using one set of variables for the entire test file, and
insted chops it up into smaller chunks, with (duplicated) sets of
those variables in each chunk.

This is needed after the merge of the set operator work[^1], as
it triggers a SPESH bug[^2] in MoarVM that crashes the test run.

[1] https://github.com/rakudo/rakudo/pull/934
[2] https://github.com/MoarVM/MoarVM/issues/446