Add more Range ~~ Range tests
Specs semantics of numeric/string Range smartmatching as discussed
at https://irclog.perlgeek.de/perl6-dev/2017-02-08#i_14064592
and https://github.com/rakudo/rakudo/pull/1017

For the 4 deleted tests: first three are subsumed by this commit,
the last one was an incorrect test. All of them were added today
and are not part of 6.c-errata.

P.S.: the fudger hangs when I try to fudge these, so I fudged manually
with a skip() and by wrapping the tests in a sub {}; remove both when
tests start to pass.