Fix broken test in S09-typed-arrays/hashes.t
The test mistakenly tests `a` key that already exists in %h prior
to type-check tests being run. The key that is being auto-vivified
in is `z`, not `a`, so change test to test for `z`'s auto-vivification.

The only reason this test was passing in the past is because failed
type check nulled the array, so the previously-present `a` key was
removed. This was fixed[^1] as part of RT130870[^2], and tests fixed[^3]
in master after discussion[^4][^5].

[1] https://github.com/rakudo/rakudo/commit/ae7bcf1b8e1fa612413dc06c
[2] https://rt.perl.org/Ticket/Display.html?id=130870
[3] https://github.com/perl6/roast/commit/7996c41781ff208b9cabf1ba
[4] https://irclog.perlgeek.de/perl6-dev/2017-02-27#i_14174359
[5] https://irclog.perlgeek.de/perl6-dev/2017-02-28#i_14175622