[io grant] Remove two fudged &chdir tests
They were fudged 5 years ago. Seems the intent is to check chdir for
non-existent path fails and not throws, but one test wouldn't work now
even if it were unfudged as a sunk Failure would explode.

Testing of this functionality will be subsumed by the new tests,
to be added shorly.