fix for correct BUILD semantics of grammar
This test uses a module that lists a bunch of internal methods.  This is
suboptimal, but for now just add the methods that are called when a proper
new/BUILDALL are done.  (The old code only worked because class Grammar
bypassed all that and went straight to nqp, which in turn made it impossible
to support user grammars with their own attributes.)

(Also, the module had the bug of passing the type object where the instantiated
object was needed.)

(Also suppress the redef warning caused by the module exporting the same name.)