Change spec of .chomp/.nl-in behaviour on cathandle
While original idea of leaving the already opened handles with
whatever mode and attributes they were in seemed neat, we run into
issues with binary/non-binary operations that might cross from
a binary handle to non-binary handle.

Spec the default to be: change mode/attrs of all handles to what the
cathandle was instantiated with. We still have room to add an option
to swap to preserving behaviour if we ever end up needing it.