Fudge new test for release
The test passes, but not when running via release bot. Probably something
to do with how $*EXECUTABLE is set or something. I changed the command
executed and I see :cwd works fine. So it's something with the test
itself.

The test is new, added just a few days ago, and isn't part of 6.c