Add new test S15-nfg/concat-stable.t
This test checks that Unicode text of different languages is able to be
broken up into the individual graphemes and concat'ed together without
any change or alteration.

The script was originally created when a bug was found in concatting things
with \r\n line endings and certain Unicode characters. This script tests
both \r\n and \n line endings and makes sure the text is stable even
at every single point in the string when concat'ed back together again.

Tests Arabic, Hangul (Korean), and Tibetan currently. The text samples are
from wikipedia and have been placed in 3rdparty/wikipedia along with LICENSE
which describes the origin pages and the license of the text
(Creative Commons Share Alike 3.0).