Add concatenation tests for NFC to ensure concat is safe.
This commit was prompted by the improvements in logic when concatenating
strings in MoarVM. The logic on whether a string needs to be
renormalized when concat'd was improved. These tests were added to ensure
that the improvements did not cause any breakage with things that change
under NFC.

Luckily the MoarVM changes pass all the new tests. These tests will be useful
to have as concatenation is further worked on.