Remove final bogus tests for bags and mixes
The referenced ticket RT #122810 was from a time when you could actually
try to build a bag from a lazy Iterable.  So whatever we were testing
before, it wasn't getting tested now, as it only generated Failures.
Since the test was really only a "lives_ok" test, and it kept living
for the wrong reason, I think it is safe to remove the test.

I have no idea how to actually test for a regression of #RT 122810.
If a test can be devised, it should live in S02-types/(bag|mix).t .