Fix is_run tests not catching errors to run the test at all
Status code is only checked if explicitly requested. So tests that just check
output on STDOUT or STDERR like "no error message was printed" won't notice
failures to run the program at all.

This also uncovers that a Perl 6 program with a MAIN exits with exit code 1
even when explicitly called with --help which looks like a mistake.