Fix S11-modules/perl6lib disturbing other tests by shadowing Test.pm6
There are quite a few spec tests that run with '.' in their module search
paths. If they try to load Test.pm6 while perl6lib.t was running, they could
end up loading the non-functional one and fail. Use a very unique name for
the test module instead.

Unfortunately we cannot just put the module into a temporary directory because
the test harness runs spec tests with ./perl6 and this will only run in the
build directory itself and the test is about the CWD.