Remove tests with some odd exepctations
They claimed to cover RT #123477, and there are indeed tests in this
file that cover that issue. However, these tests went further: they
expected not only that the protocol would be enforced from the point
of view of an individual tap (which is what the RT ticket was using),
but also that this would carry over to future taps of the same Supply.
This is not a desirable behavior for on-demand supplies at all, and
is being rectified in Rakudo.

Providing this led to a curious behavior: calling `$supplier.Supply`
once and tapping it twice in succession would behave differently to
calling it twice and tapping each result once. This doesn't seem like
a desirable difference, and certainly not worth adding extra overhead
to enforce just for live supplies, or maybe just for Supplier.

So, get rid of the tests requiring this odd behavior.