Update uniname tests to test hex after non-unique's
Instead of just testing <Control-0000> we also test for CJK
Ideographs like <CJK Ideograph-XXXX>. All names starting with a
'<' and ending with '>' are inherently nonunique names, so we
give them names similar to how they are refered to within the
Unicode dataset.

Also test that noncharacters return <noncharacter-XXXX> rather than
illegal or reserved. Check that we get reserved instead of illegal
back for characters that are non-illegal but under 0x10FFFF