Fix and unfudge incorrect Pair binding test
The current version attempts to bind a Str variable to a Signature
with one named arg in it. The test is fudged, as :$a gets a Mu
value instead of $b's value, and it doesn't outright die only
because Str.Capture returns an empty capture. In addition, if
actual bind is wanted, the :$a needs to be received raw.

Fix by binding a proper Capture, with $b passed as a named :a arg.