Adjust broken Seq consumation tests
The tests are not part of 6.c-errata.

The tests passed for the wrong reason. `eqv` will actually
return True if the LHS and RHS are the same object; the test was
passing because the Seq was touched in the guts of cmp-ok.

Fix by using different consumed Seqs in the test.