Rewrite bogus-ish dir.head.dirname test
The routine returns IO::Path objects which are composed of $!CWD
and $!path. These objects can be made cheaper by setting $!CWD to
the dir being dir()ed and $!path to the found stuff--without having
to resolve anything. For this reason, it doesn't make much sense
to require .dirname to be set to absolute paths.

Change the test to instead test that the generated IO::Paths actually
point to proper filesystem entities.