Fix broken Proc piping test
The test is not part of 6.c-errata.

.close on a pipe returns the Proc and sunk Proc awaits the
process to finish. This causes proc1 in the test to finish off
before the second proc had a chance to grab its pipe.

The only reason this test was passing in the past is due to a bug where
only the .close on the last open handle actually returned a Proc