Change buggy code in .subst in actions test
The test covers RT#71514, where the tested code uses .subst in
a context where we already have a $/ in use for other purpose.
It appears in the past such code silently did Wrong Things™ which
was fixed by adding a `try` into .subst.

However, this fix still does Wrong Things™ and the test's claim
that .subst can be used "without trouble" is incorrect. If $/
is used in the substitution portion, it will reffer to incorrect
value:

  grammar { token TOP { (.+) } }.parse("Bar", :actions(
      class { method TOP($/) { put "Foo".subst: /(Foo)/, {uc $/[0] } } }
  )); # EXPECTED: "FOO"; OUTPUT: "BAR"

The correct behaviour is to complain about $/. Rakudo HEAD will do
that shortly and roast master will cover this behaviour.