Use correcter tests for Num.Rat conversions
The test descriptions say `==` is being checked, but the `is` routine
is actually testing Str views of Rats and Nums. Change to `cmp-ok` with
`==` as comparator, since IEEE-style zero-denominator Rats are explosive
in Str views, while non-explosive in Num views, so they will match here.

Since NaN does not `==` to itself, use `.isNaN` test for it. Fudge it for now,
since there's currently a to-be-fixed bug where NaN.Rat is not `isNaN`.

More explanation on why this is being changed in v6.c-errata:

Those tests are incorrect. They're using string equality (due to use
of `is` routine) to test numeric equality of two objects. Str view of
Rat involves division and only in Num view division by zero is not an
explosion. As for the change itself. In 2017.12, we had NaN.Rat return
a special non-Rat object that knew how to convert back to NaN and did
not explode with stringification. Now, we simply return an effective
inverse operation: <0/0>.Num => NaN; NaN.Rat => <0/0>. This fixed 3
bugs with the old system and the original special-object thing was a
misinterpretation. Full description here:

https://github.com/rakudo/rakudo/commit/042cb7413f87277eadc4fa220f1a6bec1cd3dfdc