[v6.d REVIEW] Use correcter tests for Num.Rat conversions
The test descriptions say `==` is being checked, but the `is` routine
is actually testing Str views of Rats and Nums. Change to `cmp-ok` with
`==` as comparator, since IEEE-style zero-denominator Rats are explosive
in Str views, while non-explosive in Num views, so they will match here.

Since NaN does not `==` to itself, use `.isNaN` test for it. Fudge it for now,
since there's currently a to-be-fixed bug where NaN.Rat is not `isNaN`.