Change incorrect &list(Seq) tests
Per https://github.com/rakudo/rakudo/issues/1344 , a Seq is not
a "list", because it does not do a Positional role.

Looks like the original tests[^2] did start off testing for List and
because they were added alongside `cache` listop, they tested for
whether caching did not occur. Then, they were changed[^2] to test for
a Seq return value. My guess that was semi-accidental and the proper
course was meant to be to fix &list(Seq:D) to give a List.

[1] https://github.com/perl6/roast/commit/a046d7448847d327e09bc542dd
[2] https://github.com/perl6/roast/commit/5463a12cbe6dacaaef898d28d3