[v6.d REVIEW] Unfudge min/max flattening tests
The skip-fudge expects no flattening to happen with single-arg
`[min]`/`[max]` with Range arg, but I don't see what would be a
viable answer in such a case, other than min/maxing the Range's
endpoints, which would merely be *an optimization*, not a functional
difference, and IMO should not be specced. (N.B.: don't see
an easy way to perform that opt in Rakudo, as &METAOP_REDUCE_LIST
makes a routine with a +slurpy).

Change the fudge to more-than-one arg tests, to ensure flattening
is not done there, as in these cases the result would be different
if flattening were done.