Adjust throw-in-read-in-a-mode test
The test is not[^1] part of 6.c-errata.

We don't actually bother to read if we reached EOF, which is the case
with empty file the test uses, so no throwage would occur just Nil/empty buf
would be returned (which I'm personally fine with).

Adjust test to use .slurp for reading which still does throw in :a open mode.

[1] https://github.com/perl6/roast/commit/091931a717aa0507d04fcef10b81