Adjust wrong-ish exception bug cover test
The bug it covers is for an internal exception produced by 2011's
compiler. Despite the exception thrown being an X::Method::NotFound,
it contains wrong properties, such as an empty method name.

After work done to simply `.=` method call to not go through
`dispatch:<.=>` method, we now throw a textually more accurate
exception that does name the method correctly. However, this exception
is thrown from the guts of BOOTSTRAP and so it is currently an X::AdHoc,
not an X::Method::NotFound.

Modify the test to test for generic Exception type and to check message
mentions the method name.