Test segfaults with reserved characters
These tests take just ≈0.7s to run but at the same time have a fairly
high “success” rate (≈35% chance of failing on affected
commits). Don't know if it's the best way to test a flapper like this,
but I guess it's better than nothing.

Interestingly, the whole file can sometimes segfault because it
contains a reserved character and acts as a test by itself.

This resolves rakudo/rakudo#1247, but for more thorough testing
see rakudo/rakudo#1316.