Fix supporting test code
`take` returns the thing it took, which in this case is the Seq
object. Since it's a `for` loop and we do the "sink last statement"
thing in them, that Seq ends up being sunk, which causes throwage
about already-iterated Seq when we actually try to compare the result.

The only reason this wasn't throwing before is because we do the
"sink last statement" thing for methods only, not for subs (so
if author of tests had used `.take` they'd get throwage). And I think
that behaviour is a bug.

Fix the test by ensuring the `take` isn't the last statement. I confirmed
with commitable bot that the test still covers the original bug.