Loosen Num-in-sink warning text message match test
Mostly to make master roast pass with Grisu3 Num stringification
on MoarVM. After the change, the value comes out as
`6.0221408999999994e+23`. My C-only impl of the Grisu3 algo doesn't
have this issue, and I suspect the cause is the Num parsing bug
and/or nqp::div_In issue with denormals is to blame for this.

Since the test merely checks a warning is emitted for a Num in sink
context, IMO, this change to the test is OK to make. I plan to
address the aforementioned Num bugs shortly, and I forsee this
commit will be good to revert soon. If I don't make it in time for
release, IMO, the change is OK to cherry-pick to 6.c-errata