[CaR Grant] Change spec of zero-denominator rationals normalization
I deem lack of normalization a bug, and thus the tests wrong.

Since they're numerically equivalent, the .numerator/.denominator should
be the same and we should normalize them right away. This lets us also use
simpler logic in things like == and === and to have .WHICH give the same thing.

This was actually the spec before June 8, 2015, but then was changed[^1]
to expect no normalization. Neither the commit, nor discussions around
the time when the change happened indicate why it was made, so I assume
the tests were changed at the time to simply match rakudo's behaviour.

[1] https://github.com/perl6/roast/commit/60dd9f82a00ee649ab87d55ef9
[2] https://irclog.perlgeek.de/perl6/2015-06-08#i_10714325