Revert "[CaR Grant] Change spec of zero-denominator rationals normalization"
This reverts commit 798b571f5f293ec8906aad2f7256ccf53ef40656.

There are a couple more tests that expect lack of normalization and
normalizing ZDRs leads to poorer error reporting.

An attempt will be made to avoid ZDR normalization and instead to
extract ZDR's functionality into a separate role, and then we can
be a bit more liberal with on-demand, internal-only normalization.