Remove tests for deprecations that won't happen
It's become clear to me that it's going to be very difficult to add
deprecations to code that has already been codegenned.

If you want to deprecate a class or attribute, and there is a method
for it written explicitely, then you should be adding the "is DEPRECATED"
to the method.  Only allow "is DEPRECATED" on classes that do not have
a specific .new, or on attributes that don't have a explicit method for themt.