[CaR Grant] Fix incorrect typed constant test
The test declares a constant of type `Num` but tries to
assign a value of type `Int` to it. The only reason this test
didn't explode in the past was because type constraints on
constants weren't plugged into anything and were entirely ignored.

Fix test by using correct type of the value. Extra 6.d tests
in S04-declarations/constant-6.d.t (master branch) also test
the typecheck failure when types do not match.