Test that bitwise string ops maintain normalization
This adds tests for MoarVM/MoarVM#867. I fixed this in MoarVM recently
and had already added tests to make sure that synthetics were handled
properly and the bitwise ops operated on a codepoint basis.

I constructed codepoints which result in a codepoint that is normalized
to multiple codepoints: for example 2940 +& 2910 = 2908. But
2908.chr.ords will give (2849, 2876). Test AND, OR and XOR string ops.